2012-08-25 v0.13 beta
On demand fetching of lower levels of property tree. If you set max_depth to 1
each new level will be fetched from the debugging engine.
Using AutoScroll=false on forms so that ClientWidth/ClientHeight is used and
forms look better on Windows 7.
Upgraded to JVCL 3.45.
And added fix for http://issuetracker.delphi-jedi.org/view.php?id=5023
Checking to see if dbgp source temp file changed and not rewriting it.


git-svn-id: http://svn.mtk.lan/svn/monotek/users/zobo/dbgpPlugin@3392 822f994f-2fd5-0310-987d-8a363e6b5143